Correcciones, links a cuentas, contador de movimientos. V1 Beta

+++ database/migrations/2026_09_10_114934_add_is_admin_to_users_table.php
@@ -8,21 +8,29 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'is_admin')) {
 
-            $table->boolean('is_admin')
-                ->default(false)
-                ->after('password');
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->boolean('is_admin')
+                    ->default(false)
+                    ->after('password');
+
+            });
+
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'is_admin')) {
 
-            $table->dropColumn('is_admin');
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->dropColumn('is_admin');
+
+            });
+
+        }
     }
 };
