+++ database/migrations/2026_09_10_115050_add_username_to_users_table.php
@@ -8,21 +8,29 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (!Schema::hasColumn('users', 'username')) {
 
-            $table->string('username')
-                ->unique()
-                ->after('name');
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->string('username')
+                    ->unique()
+                    ->after('name');
+
+            });
+
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
+        if (Schema::hasColumn('users', 'username')) {
 
-            $table->dropColumn('username');
+            Schema::table('users', function (Blueprint $table) {
 
-        });
+                $table->dropColumn('username');
+
+            });
+
+        }
     }
 };
